admin/requirements.txt
# Core dependencies
python-dotenv==1.0.0
supabase==2.3.0
boto3==1.34.0
Pillow==10.2.0
pillow-heif==0.15.0

# AI and image processing
google-generativeai==0.3.2
opencv-python==4.9.0.80

# Data analysis and visualization
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# GUI and interface
tkinter  # Built-in with Python
requests==2.31.0

# Development and testing
pytest==7.4.3
black==23.12.0
flake8==6.1.0

# Optional: Advanced image processing
scikit-image==0.22.0
imageio==2.33.1